fix some styles, make image selector overlay rtl ready

+++ library/servers/class_upfront_javascript_main.php
@@ -337,7 +337,7 @@ class Upfront_JavascriptMain extends Upfront_Server {
 		);
 
 		$menus = json_encode(wp_get_nav_menus());
-		$is_rtl = is_rtl();
+		$is_rtl = (int) is_rtl();
 		$main = <<<EOMainJs
 // Set up the global namespace
 var Upfront = window.Upfront || {};
